requirements.txt
python-dotenv==0.19.1
pyTelegramBotAPI==4.0.0
